Issue 32: Adds 3 column light layout. Transitions Inspiration.vue to Index.vue to show as main page
